pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "franca"
version = "1.0.0"
description = "Franca: Nested Matryoshka Clustering for Scalable Visual Representation Learning (official repository)."
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Shashanka Venkataramanan", email = "shashanka.venkataramanan@valeo.com" },
    { name = "Valentinos Pariza", email = "valentinospariza@gmail.com" },
    { name = "Elias Ramzi", email = "elias.ramzi@valeo.com" },
    { name = "Henrik Vendelbo", email = "hvendelbo@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "torch>=2.2.0,<2.3.0",  # Use PyTorch 2.2.x which has macOS wheels
    "torchvision>=0.17.0,<0.18.0",

    "pytorch-lightning==2.4.0",
    "torchmetrics==1.4.3",

    "numpy==1.26,<2.0",
    "tqdm==4.66.5",
    "pillow>=10,<11",
    "scipy==1.14.0",
    "omegaconf==2.3.0",
    "webdataset==0.2.96",
    "sacred==0.8.7",
    "joblib==1.4.2",
]

[project.optional-dependencies]
franca = [
    "scikit-learn==1.5.1",
    "torchdata==0.11.0",
    "fvcore@git+https://github.com/facebookresearch/fvcore@f3b07ea37daff5b774702708134f7b702e2dc83e",
]
coreml = [
  # Installed only on macOS runners/dev machines:
  "coremltools==7.1; platform_system == 'Darwin'",
  "requests>=2.31.0",
]

dev = [
    "flake8",
    "flake8-annotations",
    "flake8-bugbear",
    "flake8-comprehensions",
    "flake8-simplify",
    "isort",
    "black[jupyter]",
    "ipython",
    "ipdb",
    "jupyter",
    "matplotlib",
    "seaborn",
]
torch = [
    "torch>=2.2.0,<2.3.0",  # Use PyTorch 2.2.x which has macOS wheels
    "torchvision>=0.17.0,<0.18.0",
]
# Linux/CUDA-only extras - temporarily commented out for macOS development
# cuml_linux = [
#     "cuml-cu12; platform_system == 'Linux' and platform_machine == 'x86_64'",
# ]
# xformers_linux = [
#     "xformers==0.0.27.post2; platform_system == 'Linux'",
# ]
# faiss_gpu = [
#     "faiss-gpu-cu12==1.10.0; platform_system == 'Linux' and platform_machine == 'x86_64'",
# ]

[project.urls]
Homepage = "https://github.com/valeoai/Franca"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["franca", "rasa"]

# Version management with hatchling
[tool.hatch.version]
path = "franca/__init__.py"

[tool.black]
line-length = 127
target-version = ['py39']
include = '\.py?$'

[tool.isort]
atomic = true
profile = "black"
line_length = 127
skip_gitignore = true

# Ensure uv resolves this repository's local package rather than any package named
# "franca" from PyPI that may define different extras (e.g., cuml-linux)
[tool.uv.sources]
franca = { path = ".", editable = true }

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "seaborn>=0.13.2",
]
